restructuring basketball pages
